Fix `Util.promisify is not a function` in API when workflow nodes are compressed (#28)
`c9dfa38a` introduced support for decoding compressed nodes on a workflow.
The runtime Node environment in Docker for the API however, does not
have support for `promisify` (it is in v8 onwards, not v6).

It is provided by a polyfill / shim in the `util.promisify` package that the
`kubernetes-client` also depends on. Thus, this commit switches to using
`util.promisify` to fix the API exception. If the runtime is upgraded in the
future, it should automatically use the standard library.
Again, as another direct dependency already depends on `util.promisify`,
we are not introducing any new dependency in this commit.